requirements.txt
python-dotenv==1.0.0
streamlit==1.31.1
pandas==1.5.3
numpy==1.23.5
Pillow==9.5.0
matplotlib==3.7.1
plotly==5.14.1
PyQt5==5.15.10
google-search-results==2.4.2
selenium==4.9.1
chromedriver-py==130.0.6723.44b0
beautifulsoup4==4.12.2
unicode==2.9
Unidecode==1.3.6
wordcloud==1.9.1.1
streamlit-carousel==0.0.4
google-cloud-aiplatform==1.48.0
langchain-google-genai==1.0.8
langchain==0.2.14
langchain-core==0.2.33
langchain-experimental==0.0.64
#langchain-google-vertexai==0.0.5




